testing cleanup

+++ test/erln8_tests.rb
@@ -1,7 +1,7 @@
 require 'rubygems'
 require 'bundler/setup'
 require "test/unit"
-
+require 'iniparse'
 class TestRepo
   def initialize(root, tags)
     `rm -rf #{root}`
@@ -33,6 +33,8 @@ echo \"Fake configure\"
 end
 
 class Erln8Test < Test::Unit::TestCase
+
+  @@teardown = true
   def setup
     `mkdir testconfig`
   end
@@ -45,9 +47,12 @@ class Erln8Test < Test::Unit::TestCase
   end
 
   def teardown
-    `rm -rf ./repo_a`
-    `rm -rf ./repo_b`
-    `rm -rf ./testconfig`
+    if @@teardown == true then
+      %w[a b c].each do |repo|
+        `rm -rf ./repo_#{repo}`
+      end
+      `rm -rf ./testconfig`
+    end
   end
 
   def test_init
@@ -73,18 +78,53 @@ class Erln8Test < Test::Unit::TestCase
   end
 
 
-  def test_other_clone
+  def test_ini
     TestRepo.new("repo_a", ["a","b","c"])
-    TestRepo.new("repo_c", ["d","e","f"])
     result = run_cmd "--init"
     configfile = "./testconfig/.erln8.d/config"
-    config = File.read(configfile)
-    File.open(configfile, "w") do |file|
-      file.puts config.gsub("https://github.com/erlang/otp.git","./repo_a")
-    end
+    ini = IniParse.parse(File.read(configfile))
+    assert ini["Repos"]["default"] == "https://github.com/erlang/otp.git"
+    assert init["Erln8"]["color"] == "false"
+    assert init["Erln8"]["banner"] == "false"
+  end
+
+  def test_ini
+    TestRepo.new("repo_a", ["a","b","c"])
+    result = run_cmd "--init"
+    configfile = "./testconfig/.erln8.d/config"
+    ini = IniParse.parse(File.read(configfile))
+    ini["Repos"]["default"] = "./repo_a"
+    ini.save(configfile)
+  end
+
+  def test_other_clone
+    TestRepo.new("repo_a", %w[a b c])
+    TestRepo.new("repo_b", %w[d e f])
+    result = run_cmd "--init"
+
+    configfile = "./testconfig/.erln8.d/config"
+    ini = IniParse.parse(File.read(configfile))
+    ini["Erln8"]["color"] = "false"
+    ini["Repos"]["default"] = "./repo_a"
+    ini["Repos"]["test_repo_b"] = "./repo_b"
+    ini["Repos"]["test_repo_c"] = "./repo_c"
+    #puts ini.to_ini
+    ini.save(configfile)
+
+    #@@teardown = false
     result = run_cmd "--clone default"
     assert File.exist?("./testconfig/.erln8.d/repos/default/.git")
+    result = run_cmd "--clone test_repo_b"
+    assert File.exist?("./testconfig/.erln8.d/repos/test_repo_b/.git")
+    result = run_cmd "--clone test_repo_c"
+    assert_equal("fatal: repository './repo_c' does not exist\n",result)
+    result = run_cmd "--clone test_repo_d"
+    assert_equal("ERROR: Unknown repository test_repo_d\n", result)
   end
+
+
+
+
 
   def run_cmd(cmd)
     c = "ERLN8_HOME=./testconfig ../erln8 #{cmd} 2>&1"
